% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection_manager.R
\name{connection_manager}
\alias{connection_manager}
\title{Simple Cassandra session manager}
\arguments{
\item{sc}{Spark context used for the scala context}

\item{str_cluster}{The target cluster to which to connect}

\item{cluster}{A pointer to a build cluster (java object)}

\item{keyspace}{The keyspace to which to connect}
}
\description{
Simple Cassandra session manager
}
