% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_poll.R
\name{cql_connect}
\alias{cql_connect}
\title{Using the cassandra java driver directly from scala
to use CQL to query the database}
\usage{
cql_connect(sc, cluster, keyspace)
}
\description{
Using the cassandra java driver directly from scala
to use CQL to query the database
}

