% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{cql_count}
\alias{cql_count}
\title{Dispatch a count to some Cassandra table}
\usage{
cql_count(sc, session, table)
}
\arguments{
\item{session}{The Cassandra session we will dispatch the query in}

\item{table}{The table to which to connect}
}
\description{
Dispatch a count to some Cassandra table
}
