% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_index.R
\name{cql_dist_file_index_query}
\alias{cql_dist_file_index_query}
\title{Dispatch the Distinct File Index query}
\usage{
cql_dist_file_index_query(sc, session)
}
\description{
Dispatch the Distinct File Index query
}
