% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_index.R
\name{cql_get_file_index}
\alias{cql_get_file_index}
\title{Exhaust file_index row iterator to get whole file_index}
\usage{
cql_get_file_index(sc, session)
}
\description{
Exhaust file_index row iterator to get whole file_index
}
