% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_responses.R
\name{cql_get_null_response_partition}
\alias{cql_get_null_response_partition}
\title{Get the partition of ordered telegrams for a vehicle}
\usage{
cql_get_null_response_partition(sc, session, date, vhid)
}
\description{
Get the partition of ordered telegrams for a vehicle
}
