% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassandra_table.R
\name{cql_get_obc_model}
\alias{cql_get_obc_model}
\title{Test joinWithCassandraTable}
\usage{
cql_get_obc_model(sc, keyspace, line, date, select_cols = list())
}
\description{
Test joinWithCassandraTable
}
