% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_index.R
\name{cql_iterate}
\alias{cql_iterate}
\title{Exhaust file_index row iterator to get whole file_index}
\usage{
cql_iterate(sc, iterator)
}
\description{
Exhaust file_index row iterator to get whole file_index
}
