% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cassandra_table.R
\name{cql_joinWithRTable}
\alias{cql_joinWithRTable}
\title{Test joinWithRTable}
\usage{
cql_joinWithRTable(sc, tbl_partitions, keyspace, table, select_cols = list())
}
\description{
Test joinWithRTable
}
