% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_responses.R
\name{cql_non_ack_polls}
\alias{cql_non_ack_polls}
\title{Gets all polls for a partition, statefully filters non-acknowledged polls and writes to Cassandra}
\usage{
cql_non_ack_polls(date, vhid)
}
\description{
Gets all polls for a partition, statefully filters non-acknowledged polls and writes to Cassandra
}
