% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_poll.R
\name{cql_previous_poll}
\alias{cql_previous_poll}
\title{Extract the previous poll R objects}
\usage{
cql_previous_poll(sc, session, query, str_date, str_vhid, str_loop_id, str_time)
}
\arguments{
\item{session}{The cassandra session}

\item{query}{The prepared string to submit as CQL}

\item{str_date}{Date passed as str}

\item{str_vhid}{Vehicle ID passed as str}

\item{str_loop_id}{Loop id passed as str}

\item{str_time}{R timestamp to be converted to and interpreted by java}
}
\description{
Extract the previous poll R objects
}
