% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previous_poll.R
\name{cql_previous_poll_query}
\alias{cql_previous_poll_query}
\title{Build the previous poll query}
\usage{
cql_previous_poll_query(sc, keyspace, table)
}
\arguments{
\item{keyspace}{The keyspace to which to connect}

\item{table}{The table to which to connect}
}
\description{
Build the previous poll query
}
