% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null_responses.R
\name{cql_write_non_ack_polls}
\alias{cql_write_non_ack_polls}
\title{Get the non-acknowledged messages}
\usage{
cql_write_non_ack_polls(sc, session, iterator)
}
\description{
Get the non-acknowledged messages
}
